Code-review changes
