+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/HardFork/Combinator/State/Types.hs
@@ -192,12 +192,13 @@ newtype TranslateTxOut x y = TranslateTxOut (TxOut (LedgerState x) -> TxOut (Led
 -- translated to newer eras. This function fills that hole and allows us to
 -- promote tables from one era into tables from the next era.
 --
--- TODO(jdral): this is not optimal. If either 'translateTxInWith' or
--- 'translateTxOutWith' is a no-op ('id'), mapping over the diff with those
--- functions is also equivalent to a no-op. However, we are still traversing the
--- map in both cases. If necessary for performance reasons, this code could be
--- optimised to skip the 'Map.mapKeys' step and/or 'Map.map' step if
--- 'translateTxInWith' and/or 'translateTxOutWith' are no-ops.
+-- NOTE: If either 'translateTxInWith' or 'translateTxOutWith' is a no-op ('id'),
+-- mapping over the diff with those functions is also equivalent to a
+-- no-op. However, we are still traversing the map in both cases.
+--
+-- NOTE: This function is only used on ticking, to prepend differences from
+-- previous eras, so it will be called only when crossing era boundaries,
+-- therefore the translation won't be equivalent to 'id'.
 translateLedgerTablesWith ::
      Ord (TxIn (LedgerState y))
   => TranslateLedgerTables x y
