+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Storage/LedgerDB/V2.hs
@@ -67,9 +67,6 @@ mkInitDb :: forall m blk.
             , IOLike m
             , LedgerDbSerialiseConstraints blk
             , HasHardForkHistory blk
-#if __GLASGOW_HASKELL__ < 906
-            , HasAnnTip blk
-#endif
             )
          => Complete LedgerDbArgs m blk
          -> Complete V2.LedgerDbFlavorArgs m
@@ -105,7 +102,7 @@ mkInitDb args flavArgs getBlock =
                , ldbCfg             = lgrConfig
                , ldbHasFS           = lgrHasFS
                , ldbResolveBlock    = getBlock
-               , ldbQueryBatchSize  = Nothing
+               , ldbQueryBatchSize  = lgrQueryBatchSize
                , ldbOpenHandlesLock = lock
                }
         h <- LDBHandle <$> newTVarIO (LedgerDBOpen env)
@@ -118,6 +115,7 @@ mkInitDb args flavArgs getBlock =
      , lgrHasFS
      , lgrSnapshotPolicyArgs
      , lgrTracer
+     , lgrQueryBatchSize
      , lgrRegistry
      } = args
 
@@ -190,6 +188,12 @@ mkInternals bss h = TestInternals {
                 (ldbHasFS env)
                 suff
                 st
+    , push = \st -> withRegistry $ \reg -> do
+          eFrk <- newForkerAtTarget h reg VolatileTip
+          case eFrk of
+            Left {} -> error "Unreachable, Volatile tip MUST be in LedgerDB"
+            Right frk ->
+              forkerPush frk st >> atomically (forkerCommit frk) >> forkerClose frk
     , reapplyThenPushNOW = \blk -> getEnv h $ \env -> withRegistry $ \reg -> do
           eFrk <- newForkerAtTarget h reg VolatileTip
           case eFrk of
@@ -404,7 +408,7 @@ data LedgerDBEnv m l blk = LedgerDBEnv {
   , ldbCfg             :: !(LedgerDbCfg l)
   , ldbHasFS           :: !(SomeHasFS m)
   , ldbResolveBlock    :: !(ResolveBlock m blk)
-  , ldbQueryBatchSize  :: !(Maybe Int)
+  , ldbQueryBatchSize  :: !QueryBatchSize
   , ldbOpenHandlesLock :: !(RAWLock m LDBLock)
   } deriving (Generic)
 
@@ -643,14 +647,13 @@ newForker h ldbEnv rr st = do
         foeLedgerSeq          = lseqVar
       , foeSwitchVar          = ldbSeq ldbEnv
       , foeSecurityParam      = ledgerDbCfgSecParam $ ldbCfg ldbEnv
-      , foeQueryBatchSize     = ldbQueryBatchSize ldbEnv
       , foeTracer             = tr
       , foeResourcesToRelease = toRelease
       }
     atomically $ modifyTVar (ldbForkers ldbEnv) $ Map.insert forkerKey forkerEnv
     pure $ Forker {
         forkerReadTables             = getForkerEnv1   h forkerKey implForkerReadTables
-      , forkerRangeReadTables        = getForkerEnv1   h forkerKey implForkerRangeReadTables
+      , forkerRangeReadTables        = getForkerEnv1   h forkerKey (implForkerRangeReadTables (ldbQueryBatchSize ldbEnv))
       , forkerGetLedgerState         = getForkerEnvSTM h forkerKey implForkerGetLedgerState
       , forkerReadStatistics         = getForkerEnv    h forkerKey implForkerReadStatistics
       , forkerPush                   = getForkerEnv1   h forkerKey implForkerPush
