+++ ouroboros-consensus/test/storage-test/Test/Ouroboros/Storage/ChainDB/StateMachine.hs
@@ -165,8 +165,6 @@ data Cmd blk it flr
     -- ^ Advance the current slot to the block's slot (unless smaller than the
     -- current slot), add the block and run chain selection.
   | GetCurrentChain
-  -- TODO(js_ldb): reenable
-  --  GetLedgerDB
   | GetTipBlock
   | GetTipHeader
   | GetTipPoint
@@ -385,7 +383,6 @@ run env@ChainDBEnv { varDB, .. } cmd =
     readTVarIO varDB >>= \st@ChainDBState { chainDB = ChainDB{..}, internal } -> case cmd of
       AddBlock blk             -> Point               <$> (advanceAndAdd st (blockSlot blk) blk)
       GetCurrentChain          -> Chain               <$> atomically getCurrentChain
-      -- GetLedgerDB              -> LedgerDB . flush    <$> atomically getDbChangelog -- TODO(jdral_ldb)
       GetTipBlock              -> MbBlock             <$> getTipBlock
       GetTipHeader             -> MbHeader            <$> getTipHeader
       GetTipPoint              -> Point               <$> atomically getTipPoint
@@ -449,33 +446,6 @@ run env@ChainDBEnv { varDB, .. } cmd =
     giveWithEq :: a -> m (WithEq a)
     giveWithEq a =
       fmap (`WithEq` a) $ atomically $ stateTVar varNextId $ \i -> (i, succ i)
-
--- | When the model is asked for the ledger DB, it reconstructs it by applying
--- the blocks in the current chain, starting from the initial ledger state.
--- Before the introduction of UTxO HD, this approach resulted in a ledger DB
--- equivalent to the one maintained by the SUT. However, after UTxO HD, this is
--- no longer the case since the ledger DB can be altered as the result of taking
--- snapshots or opening the ledger DB (for instance when we process the
--- 'WipeVolatileDB' command). Taking snapshots or opening the ledger DB cause
--- the ledger DB to be flushed, which modifies its sequence of volatile and
--- immutable states.
---
--- The model does not have information about when the flushes occur and it
--- cannot infer that information in a reliable way since this depends on the low
--- level details of operations such as opening the ledger DB. Therefore, we
--- assume that the 'GetLedgerDB' command should return a flushed ledger DB, and
--- we use this function to implement such command both in the SUT and in the
--- model.
---
--- When we compare the SUT and model's ledger DBs, by flushing we are not
--- comparing the immutable parts of the SUT and model's ledger DBs. However,
--- this was already the case in before the introduction of UTxO HD: if the
--- current chain contained more than K blocks, then the ledger states before the
--- immutable tip were not compared by the 'GetLedgerDB' command.
--- flush ::
---      (LedgerSupportsProtocol blk)
---   => DbChangelog.DbChangelog' blk -> DbChangelog.DbChangelog' blk
--- flush = snd . DbChangelog.splitForFlushing
 
 persistBlks :: IOLike m => ShouldGarbageCollect -> ChainDB.Internal m blk -> m ()
 persistBlks collectGarbage ChainDB.Internal{..} = do
@@ -642,7 +612,6 @@ runPure :: forall blk.
 runPure cfg = \case
     AddBlock blk             -> ok  Point               $ update  (add blk)
     GetCurrentChain          -> ok  Chain               $ query   (Model.volatileChain k getHeader)
---    GetLedgerDB              -> ok  LedgerDB            $ query   (flush . Model.getDbChangelog cfg)
     GetTipBlock              -> ok  MbBlock             $ query    Model.tipBlock
     GetTipHeader             -> ok  MbHeader            $ query   (fmap getHeader . Model.tipBlock)
     GetTipPoint              -> ok  Point               $ query    Model.tipPoint
