+++ ouroboros-consensus/test/storage-test/Test/Ouroboros/Storage/LedgerDB.hs
@@ -10,18 +10,18 @@ import qualified Test.Ouroboros.Storage.LedgerDB.Serialisation as Serialisation
 import qualified Test.Ouroboros.Storage.LedgerDB.SnapshotPolicy as SnapshotPolicy
 import qualified Test.Ouroboros.Storage.LedgerDB.StateMachine as StateMachine
 import qualified Test.Ouroboros.Storage.LedgerDB.V1.BackingStore as BackingStore
-import qualified Test.Ouroboros.Storage.LedgerDB.V1.DbChangelog.QuickCheck as DbChangelog.QuickCheck
-import qualified Test.Ouroboros.Storage.LedgerDB.V1.DbChangelog.Unit as DbChangelog.Unit
+import qualified Test.Ouroboros.Storage.LedgerDB.V1.DbChangelog as DbChangelog
 import           Test.Tasty (TestTree, testGroup)
 
 tests :: TestTree
 tests = testGroup "LedgerDB" [
       testGroup "V1" [
           BackingStore.tests
-        , DbChangelog.Unit.tests
-        , DbChangelog.QuickCheck.tests
+        , DbChangelog.tests
     ]
+      -- Independent of the LedgerDB implementation
     , SnapshotPolicy.tests
     , Serialisation.tests
+      -- Tests both V1 and V2
     , StateMachine.tests
     ]
