--- ouroboros-consensus/test/storage-test/Test/Ouroboros/Storage/LedgerDB/V1/DbChangelog/QuickCheck.hs
@@ -1,336 +1,0 @@
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE KindSignatures #-}
-{-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE UndecidableInstances #-}
-
-{-# OPTIONS_GHC -Wno-orphans #-}
-
--- | In-memory ledger DB tests.
---
--- The in-memory component of the ledger DB is a bit tricky: it stores only a
--- few snapshots of the ledger state, in order to reduce memory footprint, but
--- must nonetheless be able to construct any ledger state (within @k@ blocks
--- from the chain tip) efficiently. The properties we are verify here are
--- various invariants of this data type, things such as
---
--- * Rolling back and then reapplying the same blocks is an identity operation
---   (provided the rollback is not too far)
--- * The shape of the datatype (where we store snapshots and how many we store)
---   always matches the policy set by the user, and is invariant under any of
---   the operations (add a block, switch to a fork, etc.)
--- * The maximum rollback supported is always @k@ (unless we are near genesis)
--- * etc.
---
-module Test.Ouroboros.Storage.LedgerDB.V1.DbChangelog.QuickCheck (tests) where
-
-import           Data.Maybe (fromJust)
-import           Data.Word
-import           Ouroboros.Consensus.Block
-import           Ouroboros.Consensus.BlockchainTime
-import           Ouroboros.Consensus.Config
-import qualified Ouroboros.Consensus.HardFork.History as HardFork
-import           Ouroboros.Consensus.Ledger.Abstract
-import           Ouroboros.Consensus.Storage.LedgerDB.API
-import           Ouroboros.Consensus.Storage.LedgerDB.V1.DbChangelog hiding
-                     (tip)
-import           Ouroboros.Consensus.Util
-import           Test.QuickCheck
-import           Test.Tasty
-import           Test.Tasty.QuickCheck
-import           Test.Util.Orphans.Arbitrary ()
-import           Test.Util.QuickCheck
-import           Test.Util.TestBlock
-
-tests :: TestTree
-tests = testGroup "InMemory" [
-      testGroup "Genesis" [
-          testProperty "current"                prop_genesisCurrent
-        ]
-    , testGroup "Push" [
-          testProperty "expectedLedger"         prop_pushExpectedLedger
-        , testProperty "pastLedger"             prop_pastLedger
-        ]
-    , testGroup "Rollback" [
-          testProperty "maxRollbackGenesisZero" prop_maxRollbackGenesisZero
-        , testProperty "ledgerDbMaxRollback"    prop_snapshotsMaxRollback
-        , testProperty "switchSameChain"        prop_switchSameChain
-        , testProperty "switchExpectedLedger"   prop_switchExpectedLedger
-        , testProperty "pastAfterSwitch"        prop_pastAfterSwitch
-        ]
-    ]
-
-{-------------------------------------------------------------------------------
-  Genesis
--------------------------------------------------------------------------------}
-
-prop_genesisCurrent :: Property
-prop_genesisCurrent =
-    current genSnaps === convertMapKind testInitLedger
-  where
-    genSnaps = empty (convertMapKind testInitLedger)
-
-{-------------------------------------------------------------------------------
-  Constructing snapshots
--------------------------------------------------------------------------------}
-
-prop_pushExpectedLedger :: ChainSetup -> Property
-prop_pushExpectedLedger setup@ChainSetup{..} =
-    classify (chainSetupSaturated setup) "saturated" $
-      conjoin [
-          l === convertMapKind (refoldLedger cfg (expectedChain o) (convertMapKind testInitLedger))
-        | (o, l) <- snapshots csPushed
-        ]
-  where
-    expectedChain :: Word64 -> [TestBlock]
-    expectedChain o = take (fromIntegral (csNumBlocks - o)) csChain
-
-    cfg :: LedgerConfig TestBlock
-    cfg = ledgerDbCfg (csBlockConfig setup)
-
-prop_pastLedger :: ChainSetup -> Property
-prop_pastLedger setup@ChainSetup{..} =
-    classify (chainSetupSaturated setup) "saturated"    $
-    classify withinReach                 "within reach" $
-          getPastLedgerAt tip csPushed
-      === if withinReach
-            then Just (current afterPrefix)
-            else Nothing
-  where
-    prefix :: [TestBlock]
-    prefix = take (fromIntegral csPrefixLen) csChain
-
-    tip :: Point TestBlock
-    tip = maybe GenesisPoint blockPoint (lastMaybe prefix)
-
-    afterPrefix :: DbChangelog (LedgerState TestBlock)
-    afterPrefix = reapplyThenPushMany' (csBlockConfig setup) prefix csGenSnaps
-
-    -- See 'prop_snapshotsMaxRollback'
-    withinReach :: Bool
-    withinReach = (csNumBlocks - csPrefixLen) <= maxRollback csPushed
-
-{-------------------------------------------------------------------------------
-  Rollback
--------------------------------------------------------------------------------}
-
-prop_maxRollbackGenesisZero :: Property
-prop_maxRollbackGenesisZero =
-        maxRollback (empty (convertMapKind testInitLedger))
-    === 0
-
-prop_snapshotsMaxRollback :: ChainSetup -> Property
-prop_snapshotsMaxRollback setup@ChainSetup{..} =
-    classify (chainSetupSaturated setup) "saturated" $
-      conjoin [
-          if chainSetupSaturated setup
-            then (maxRollback csPushed) `ge` k
-            else (maxRollback csPushed) `ge` (min k csNumBlocks)
-        , (maxRollback csPushed) `le` k
-        ]
-  where
-    SecurityParam k = csSecParam
-
-prop_switchSameChain :: SwitchSetup -> Property
-prop_switchSameChain setup@SwitchSetup{..} =
-    classify (switchSetupSaturated setup) "saturated" $
-          switch' (csBlockConfig ssChainSetup) ssNumRollback blockInfo csPushed
-      === Just csPushed
-  where
-    ChainSetup{csPushed} = ssChainSetup
-    blockInfo            = ssRemoved
-
-prop_switchExpectedLedger :: SwitchSetup -> Property
-prop_switchExpectedLedger setup@SwitchSetup{..} =
-    classify (switchSetupSaturated setup) "saturated" $
-      conjoin [
-          l === convertMapKind (refoldLedger cfg (expectedChain o) (convertMapKind testInitLedger))
-        | (o, l) <- snapshots ssSwitched
-        ]
-  where
-    expectedChain :: Word64 -> [TestBlock]
-    expectedChain o = take (fromIntegral (ssNumBlocks - o)) ssChain
-
-    cfg :: LedgerConfig TestBlock
-    cfg = ledgerDbCfg (csBlockConfig ssChainSetup)
-
--- | Check 'prop_pastLedger' still holds after switching to a fork
-prop_pastAfterSwitch :: SwitchSetup -> Property
-prop_pastAfterSwitch setup@SwitchSetup{..} =
-    classify (switchSetupSaturated setup) "saturated"    $
-    classify withinReach                  "within reach" $
-          getPastLedgerAt tip ssSwitched
-      === if withinReach
-            then Just (current afterPrefix)
-            else Nothing
-  where
-    prefix :: [TestBlock]
-    prefix = take (fromIntegral ssPrefixLen) ssChain
-
-    tip :: Point TestBlock
-    tip = maybe GenesisPoint blockPoint (lastMaybe prefix)
-
-    afterPrefix :: DbChangelog (LedgerState TestBlock)
-    afterPrefix = reapplyThenPushMany' (csBlockConfig ssChainSetup) prefix (csGenSnaps ssChainSetup)
-
-    -- See 'prop_snapshotsMaxRollback'
-    withinReach :: Bool
-    withinReach = (ssNumBlocks - ssPrefixLen) <= maxRollback ssSwitched
-
-{-------------------------------------------------------------------------------
-  Test setup
--------------------------------------------------------------------------------}
-
-data ChainSetup = ChainSetup {
-      -- | Security parameter
-      csSecParam  :: SecurityParam
-
-      -- | Number of blocks applied
-    , csNumBlocks :: Word64
-
-      -- | Some prefix of the chain
-      --
-      -- Although we choose this to be less than or equal to 'csNumBlocks',
-      -- we don't guarantee this during shrinking. If 'csPrefixLen' is larger
-      -- than 'csNumBlocks', the prefix should simply be considered to be the
-      -- entire chain.
-    , csPrefixLen :: Word64
-
-      -- | Derived: genesis snapshots
-    , csGenSnaps  :: DbChangelog (LedgerState TestBlock)
-
-      -- | Derived: the actual blocks that got applied (old to new)
-    , csChain     :: [TestBlock]
-
-      -- | Derived: the snapshots after all blocks were applied
-    , csPushed    :: DbChangelog (LedgerState TestBlock)
-    }
-  deriving (Show)
-
-csBlockConfig :: ChainSetup -> LedgerDbCfg (LedgerState TestBlock)
-csBlockConfig = csBlockConfig' . csSecParam
-
-csBlockConfig' :: SecurityParam -> LedgerDbCfg (LedgerState TestBlock)
-csBlockConfig' secParam = LedgerDbCfg {
-      ledgerDbCfgSecParam = secParam
-    , ledgerDbCfg         =
-          testBlockLedgerConfigFrom
-        $ HardFork.defaultEraParams secParam slotLength
-    }
-  where
-    slotLength = slotLengthFromSec 20
-
-chainSetupSaturated :: ChainSetup -> Bool
-chainSetupSaturated ChainSetup{..} = isSaturated csSecParam csPushed
-
-data SwitchSetup = SwitchSetup {
-      -- | Chain setup
-      ssChainSetup  :: ChainSetup
-
-      -- | Number of blocks to roll back
-    , ssNumRollback :: Word64
-
-      -- | Number of new blocks (to be applied after the rollback)
-    , ssNumNew      :: Word64
-
-      -- | Prefix of the new chain
-      --
-      -- See also 'csPrefixLen'
-    , ssPrefixLen   :: Word64
-
-      -- | Derived: number of blocks in the new chain
-    , ssNumBlocks   :: Word64
-
-      -- | Derived: the blocks that were removed
-    , ssRemoved     :: [TestBlock]
-
-      -- | Derived: the new blocks themselves
-    , ssNewBlocks   :: [TestBlock]
-
-      -- | Derived: the full chain after switching to this fork
-    , ssChain       :: [TestBlock]
-
-      -- | Derived; the snapshots after the switch was performed
-    , ssSwitched    :: DbChangelog (LedgerState TestBlock)
-    }
-  deriving (Show)
-
-switchSetupSaturated :: SwitchSetup -> Bool
-switchSetupSaturated = chainSetupSaturated . ssChainSetup
-
-mkTestSetup :: SecurityParam -> Word64 -> Word64 -> ChainSetup
-mkTestSetup csSecParam csNumBlocks csPrefixLen =
-    ChainSetup {..}
-  where
-    csGenSnaps = empty (convertMapKind testInitLedger)
-    csChain    = take (fromIntegral csNumBlocks) $
-                   iterate successorBlock (firstBlock 0)
-    csPushed   = reapplyThenPushMany' (csBlockConfig' csSecParam) csChain csGenSnaps
-
-mkRollbackSetup :: ChainSetup -> Word64 -> Word64 -> Word64 -> SwitchSetup
-mkRollbackSetup ssChainSetup ssNumRollback ssNumNew ssPrefixLen =
-    SwitchSetup {..}
-  where
-    ChainSetup{..} = ssChainSetup
-
-    ssNumBlocks = csNumBlocks - ssNumRollback + ssNumNew
-    ssRemoved   = takeLast ssNumRollback csChain
-    ssNewBlocks = let afterRollback      = dropLast ssNumRollback csChain
-                      firstAfterRollback =
-                        case lastMaybe afterRollback of
-                          Nothing -> firstBlock 1
-                          Just b  -> modifyFork (+ 1) $ successorBlock b
-                  in take (fromIntegral ssNumNew) $
-                        iterate successorBlock firstAfterRollback
-    ssChain     = concat [
-                         take (fromIntegral (csNumBlocks - ssNumRollback)) csChain
-                       , ssNewBlocks
-                       ]
-    ssSwitched  = fromJust $ switch' (csBlockConfig ssChainSetup) ssNumRollback ssNewBlocks csPushed
-
-instance Arbitrary ChainSetup where
-  arbitrary = do
-      secParam <- arbitrary
-      let k = maxRollbacks secParam
-      numBlocks <- choose (0, k * 2)
-      prefixLen <- choose (0, numBlocks)
-      return $ mkTestSetup secParam numBlocks prefixLen
-
-  shrink ChainSetup{..} = concat [
-        -- Shrink the policy
-        [ mkTestSetup csSecParam' csNumBlocks csPrefixLen
-        | csSecParam' <- shrink csSecParam
-        ]
-
-        -- Reduce number of blocks
-      , [ mkTestSetup csSecParam csNumBlocks' csPrefixLen
-        | csNumBlocks' <- shrink csNumBlocks
-        ]
-      ]
-
-instance Arbitrary SwitchSetup where
-  arbitrary = do
-      chainSetup  <- arbitrary
-      numRollback <- choose (0, maxRollback (csPushed chainSetup))
-      numNew      <- choose (numRollback, 2 * numRollback)
-      prefixLen   <- choose (0, csNumBlocks chainSetup - numRollback + numNew)
-      return $ mkRollbackSetup chainSetup numRollback numNew prefixLen
-
-  shrink SwitchSetup{..} = concat [
-        -- If we shrink the chain setup, we might restrict max rollback
-        [ mkRollbackSetup ssChainSetup' ssNumRollback ssNumNew ssPrefixLen
-        | ssChainSetup' <- shrink ssChainSetup
-        , ssNumRollback <= maxRollback (csPushed ssChainSetup')
-        ]
-        -- Number of new blocks must be at least the rollback
-      , [ mkRollbackSetup ssChainSetup ssNumRollback ssNumNew' ssPrefixLen
-        | ssNumNew' <- shrink ssNumNew
-        , ssNumNew' >= ssNumRollback
-        ]
-        -- But rolling back less is always possible
-      , [ mkRollbackSetup ssChainSetup ssNumRollback' ssNumNew ssPrefixLen
-        | ssNumRollback' <- shrink ssNumRollback
-        ]
-      ]
